Lift ViewModel out of AppInfo.kt Keep everything nice and separated, we use callbacks instead
